Create CloudImage.js
CloudImage makes it easy to "set it and forget" an image that your pulling from an asynchronous source (e.g. Appcelerator Cloud) - this module wraps the ImageView and lets you pass in a Photo id from ACS and this module handles the rest - no need to capture the callback and find your image in a tableview and reset it etc. Hope it helps you out.